修bug4

+++ src/main/java/com/example/demo/config/securityConfig/provider/PhoneAuthenticationProvider.java
@@ -1,5 +1,7 @@
 package com.example.demo.config.securityConfig.provider;
 
+import com.example.demo.config.exception.AppException;
+import com.example.demo.config.exception.AppExceptionCodeMsg;
 import com.example.demo.config.securityConfig.token.PhoneAuthenticationToken;
 import com.example.demo.config.securityConfig.token.WeChatAuthenticationToken;
 import com.example.demo.entity.User;
@@ -7,6 +9,7 @@ import com.example.demo.mapper.UserMapper;
 import com.example.demo.utils.RedisCache;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.GrantedAuthority;
@@ -28,8 +31,11 @@ public class PhoneAuthenticationProvider implements AuthenticationProvider {
         PhoneAuthenticationToken authenticationToken=(PhoneAuthenticationToken)authentication;
         String phone=(String)authenticationToken.getPrincipal();//获取电话
         String code=(String)authenticationToken.getCredentials();//获取验证码
-        //todo 验证码校验，然后返回数据库信息
+
        String code1= (String)redisCache.getCacheObject("Phone_"+phone);
+       if(code1==null){
+           throw new BadCredentialsException("验证码无效");
+       }
       if(code1.equals(code)){
           redisCache.deleteObject("Phone_"+phone);
           Map map=new HashMap();
@@ -37,7 +43,8 @@ public class PhoneAuthenticationProvider implements AuthenticationProvider {
           List<User> list=userMapper.selectByMap(map);
           Set permissionSet=new HashSet();
           if(list.size()==0){
-              //todo 有问题，该手机号没有用户
+
+              throw new BadCredentialsException("该手机号没有用户");
           }
           permissionSet = userMapper.selectPermissionByUserId(list.get(0).getUserId());
           List<GrantedAuthority> authorities = new ArrayList<>();
